requirements.txt
# FocalCodec 25Hz 2048 Training Requirements
# Python 3.9+ recommended

# Core
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0

# Hugging Face
huggingface-hub>=0.16.0
safetensors>=0.3.0
transformers>=4.30.0

# Audio Processing
soundfile>=0.12.0
librosa>=0.10.0

# ASR for Evaluation
openai-whisper>=20231117

# Metrics
Levenshtein>=0.21.0

# Optional: Stage 2/3 Training
# speechbrain>=0.5.0  # For HuBERT
